Add comments and delete some useless methods.
