+++ Assets/WillDelete/Editor/view/MenuMethod.cs
@@ -7,7 +7,7 @@ public class MenuMethod : EditorWindow {
 	private static EditorWindow _window;
 	[MenuItem("CrevoxExtend/Volume generation", false, 1)]
 	public static void ShowWindow() {
-		_window = EditorWindow.GetWindow<CrevoxExtend.VolumeDataTransformWindow>("Volume generation", true);
+		_window = EditorWindow.GetWindow<CrevoxExtend.CrevoxGenerationWindow>("Volume generation", true);
 		_window.position = new Rect(35, 35, 300, 50);
 	}
 	[MenuItem("CrevoxExtend/SpaceAlphabet", false, 1)]
